+++ src/main/java/team3176/robot/subsystems/superstructure/climb/Climb.java
@@ -1,4 +1,4 @@
-package team3176.robot.subsystems.superstructure;
+package team3176.robot.subsystems.superstructure.climb;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import org.littletonrobotics.junction.Logger;
