+++ src/main/java/team3176/robot/subsystems/superstructure/climb/ClimbIOSim.java
@@ -5,7 +5,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package team3176.robot.subsystems.superstructure;
+package team3176.robot.subsystems.superstructure.climb;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.system.plant.DCMotor;
